Update disk_core3.py
A code for generating an image with computed path was added.
Also, after the algorithm fails to find a suitable path, it doesnt crash, but rather prints that path could not be found and starts anew. I did not figure out how to stop the whole cycle reliably, so it just announces that path was not found and keeps trying to find it indefinitely.